requirements.txt
fastapi==0.115.8
starlette==0.45.3
uvicorn==0.34.0
SQLAlchemy==2.0.38
pydantic==2.10.6
PyJWT==2.10.1
openpyxl==3.1.5
python-multipart==0.0.20
python-dotenv==1.0.1
psycopg[binary]==3.2.6
